evaluate tensordot performance. esp on 1x1 conv2d layers
test_conv2d generates a relay mod with a single conv2d operator. The CMSIS-NN still performs better then the tensordot for layers with many output channels. I believe it has something to do with the Flash memory accesses. I tried to reorder instrunctions and improve the performance of native schedule, it was reduced from 31ms to 20ms but still not as good as CMSIS-NN at 13ms.